refactor(app): split classic edit shell
